refactor: moved db upgrade to separate post build script
